add - delete offer&purchase

+++ InvoiceingProduct/InvoiceingProduct/Controllers/OfferController.cs
@@ -1,24 +1,29 @@
 ﻿using InvoiceingProduct.Data;
 using InvoiceingProduct.Models;
 using InvoiceingProduct.Repository;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace InvoiceingProduct.Controllers
 {
+    [Authorize(Roles = "Accountant,Purchaser,Admin")]
     public class OfferController : Controller
     {
         private OfferRepository _offerRepository;
         private ProductRepository _productRepository;
         private VendorRepository _vendorRepository;
+        private PurchaseRepository _purchaseRepository;
 
         public OfferController(ApplicationDbContext dbcontext)
         {
             _productRepository = new ProductRepository(dbcontext);
             _vendorRepository = new VendorRepository(dbcontext);
             _offerRepository = new OfferRepository(dbcontext);
+            _purchaseRepository = new PurchaseRepository(dbcontext);
         }
+        
         // GET: OfferController
         public ActionResult Index()
         {
@@ -34,8 +39,10 @@ namespace InvoiceingProduct.Controllers
 
             return View(list);
         }
+       
 
         // GET: OfferController/Details/5
+
         public ActionResult Details(Guid id)
         {
             var model = _offerRepository.GetOfferById(id);
@@ -48,6 +55,7 @@ namespace InvoiceingProduct.Controllers
         }
 
         // GET: OfferController/Create
+        [Authorize(Roles = "Purchaser,Admin")]
         public ActionResult Create()
         {
             var products = _productRepository.GetAllProducts();
@@ -60,6 +68,7 @@ namespace InvoiceingProduct.Controllers
         }
 
         // POST: OfferController/Create
+        [Authorize(Roles = "Purchaser,Admin")]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Create(IFormCollection collection)
@@ -83,6 +92,7 @@ namespace InvoiceingProduct.Controllers
         }
 
         // GET: OfferController/Edit/5
+        [Authorize(Roles = "Purchaser,Admin")]
         public ActionResult Edit(Guid id)
         {
             var model = _offerRepository.GetOfferById(id);
@@ -92,13 +102,12 @@ namespace InvoiceingProduct.Controllers
             var vendors = _vendorRepository.GetAllVendors();
             var vendorList = vendors.Select(x => new SelectListItem() { Text = x.Name, Value = x.IdVendor.ToString() });
             ViewBag.VendorList = vendorList;
-           
-           
-
+ 
             return View("EditOffer",model);
-                   }
+         }
 
         // POST: OfferController/Edit/5
+        [Authorize(Roles = "Purchaser,Admin")]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Edit(Guid id, IFormCollection collection)
@@ -125,8 +134,10 @@ namespace InvoiceingProduct.Controllers
         }
 
         // GET: OfferController/Delete/5
+        [Authorize(Roles = "Purchaser,Admin")]
         public ActionResult Delete(Guid id)
         {
+            ViewBag.ErrorMessage = TempData["OfferErrorMessage"];
             var model = _offerRepository.GetOfferById(id);
             var product = _productRepository.GetProductById(model.IdProduct);
             model.ProductName = product.ProductName;
@@ -136,17 +147,39 @@ namespace InvoiceingProduct.Controllers
         }
 
         // POST: OfferController/Delete/5
+        [Authorize(Roles = "Purchaser,Admin")]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Delete(Guid id, IFormCollection collection)
         {
             try
             {
-                _offerRepository.DeleteOffer(id);
-                return RedirectToAction(nameof(Index));
+                var listPurchase = _purchaseRepository.GetAllPurchases();
+                bool hasPurchase = false;
+                foreach (var purchase in listPurchase)
+                {
+                    if (purchase.IdOffer == id)
+                    {
+                        hasPurchase=true;
+                        break;
+                    }
+                }
+
+                if (hasPurchase == false)
+                {
+                    _offerRepository.DeleteOffer(id);
+                     return RedirectToAction(nameof(Index));
+                }
+                else
+                {
+                    TempData["OfferErrorMessage"] = "Cannot delete offer because has purchase associated.";
+                    return RedirectToAction("Delete", id);
+                }
+
             }
             catch(Exception e)
             {
+                TempData["OfferErrorMessage"] = e.Message;
                 return RedirectToAction("Delete",id);
             }
         }
