+++ InvoiceingProduct/InvoiceingProduct/Controllers/ProductController.cs
@@ -1,17 +1,23 @@
 ﻿using InvoiceingProduct.Data;
 using InvoiceingProduct.Models;
 using InvoiceingProduct.Repository;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace InvoiceingProduct.Controllers
 {
+    [Authorize(Roles = "Accountant,Purchaser,Admin")]
     public class ProductController : Controller
     {
         private ProductRepository _productRepository;
+        private OfferRepository _offerRepository;
+        private PurchaseRepository _purchaseRepository; 
         public ProductController(ApplicationDbContext dbcontext)
         {
             _productRepository = new ProductRepository(dbcontext);
+            _offerRepository = new OfferRepository(dbcontext);
+            _purchaseRepository = new PurchaseRepository(dbcontext);
         }
         // GET: ProductController
         public ActionResult Index()
@@ -28,12 +34,14 @@ namespace InvoiceingProduct.Controllers
         }
 
         // GET: ProductController/Create
+        [Authorize(Roles = "Purchaser,Admin")]
         public ActionResult Create()
         {
             return View("CreateProduct");
         }
 
         // POST: ProductController/Create
+        [Authorize(Roles = "Purchaser,Admin")]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Create(IFormCollection collection)
@@ -56,6 +64,7 @@ namespace InvoiceingProduct.Controllers
         }
 
         // GET: ProductController/Edit/5
+        [Authorize(Roles = "Purchaser,Admin")]
         public ActionResult Edit(Guid id)
         {
             var model = _productRepository.GetProductById(id);
@@ -63,6 +72,7 @@ namespace InvoiceingProduct.Controllers
         }
 
         // POST: ProductController/Edit/5
+        [Authorize(Roles = "Purchaser,Admin")]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Edit(Guid id, IFormCollection collection)
@@ -89,24 +99,68 @@ namespace InvoiceingProduct.Controllers
         }
 
         // GET: ProductController/Delete/5
+        [Authorize(Roles = "Purchaser,Admin")]
         public ActionResult Delete(Guid id)
         {
+            ViewBag.ErrorMessage = TempData["ProductErrorMessage"]?.ToString();
             var model = _productRepository.GetProductById(id);
             return View("DeleteProduct",model);
         }
 
         // POST: ProductController/Delete/5
+        [Authorize(Roles = "Purchaser,Admin")]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Delete(Guid id, IFormCollection collection)
         {
             try
             {
-                _productRepository.DeleteProduct(id);
-                return RedirectToAction(nameof(Index));
+                var listOffer = _offerRepository.GetAllOffers();
+                var listPurchase = _purchaseRepository.GetAllPurchases();
+                bool hasOffer =false;
+                bool hasPurchase = false;
+
+                foreach(var offer in listOffer)
+                {
+                    if(offer.IdProduct == id)
+                    {
+                        hasOffer = true;
+                        foreach(var purchase in listPurchase)
+                        {
+                            if(purchase.IdOffer == offer.IdOffer)
+                            {
+                                hasPurchase = true;
+                                break;
+                            }
+                        }
+                    }
+                }
+
+                if (!hasPurchase)
+                {
+                    if (hasOffer)
+                    {
+                        foreach (var offer in listOffer)
+                        {
+                            if (offer.IdProduct == id)
+                            {
+                                _offerRepository.DeleteOffer(offer.IdOffer);
+                            }
+                        }
+                    }
+                    _productRepository.DeleteProduct(id);
+
+                    return RedirectToAction(nameof(Index));
+                }
+                else
+                {
+                    TempData["ProductErrorMessage"] = "This product is associated with an offer that has a purchase. Cannot delete!";
+                    return RedirectToAction("Delete", id);
+                }
             }
             catch(Exception ex)
             {
+                TempData["ProductErrorMessage"] = ex.Message;
                 return RedirectToAction("Delete",id);
             }
         }
