+++ InvoiceingProduct/InvoiceingProduct/Controllers/VendorController.cs
@@ -1,17 +1,23 @@
 ﻿using InvoiceingProduct.Data;
 using InvoiceingProduct.Models;
 using InvoiceingProduct.Repository;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace InvoiceingProduct.Controllers
 {
+    [Authorize(Roles = "Purchaser,Accountant,Admin")]
     public class VendorController : Controller
     {
         private VendorRepository _vendorRepository;
+        private OfferRepository _offerRepository;
+        private PurchaseRepository _purchaseRepository;
         public VendorController(ApplicationDbContext dbcontext)
         {
             _vendorRepository = new VendorRepository(dbcontext);
+            _offerRepository = new OfferRepository(dbcontext);
+            _purchaseRepository = new PurchaseRepository(dbcontext);
         }
         public ActionResult Index()
         {
@@ -27,12 +33,14 @@ namespace InvoiceingProduct.Controllers
         }
 
         // GET: VendorController/Create
+        [Authorize(Roles = "Purchaser")]
         public ActionResult Create()
         {
             return View("CreateVendor");
         }
 
         // POST: VendorController/Create
+        [Authorize(Roles = "Purchaser,Admin")]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Create(IFormCollection collection)
@@ -55,6 +63,7 @@ namespace InvoiceingProduct.Controllers
         }
 
         // GET: VendorController/Edit/5
+        [Authorize(Roles = "Purchaser,Admin")]
         public ActionResult Edit(Guid id)
         {
             var model = _vendorRepository.GetVendorById(id);
@@ -62,6 +71,7 @@ namespace InvoiceingProduct.Controllers
         }
 
         // POST: VendorController/Edit/5
+        [Authorize(Roles = "Purchaser,Admin")]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Edit(Guid id, IFormCollection collection)
@@ -87,24 +97,67 @@ namespace InvoiceingProduct.Controllers
         }
 
         // GET: VendorController/Delete/5
+        [Authorize(Roles = "Purchaser,Admin")]
         public ActionResult Delete(Guid id)
         {
+            ViewBag.ErrorMessage = TempData["VendorErrorMessage"];
             var model = _vendorRepository.GetVendorById(id);
             return View("DeleteVendor",model);
         }
 
         // POST: VendorController/Delete/5
+        [Authorize(Roles = "Purchaser,Admin")]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Delete(Guid id, IFormCollection collection)
         {
             try
             {
-                _vendorRepository.DeleteVendor(id);
-                return RedirectToAction(nameof(Index));
+                var listOffer = _offerRepository.GetAllOffers();
+                var listPurchase = _purchaseRepository.GetAllPurchases();
+                bool hasOffer = false;
+                bool hasPurchase = false;
+
+                foreach (var offer in listOffer)
+                {
+                    if(offer.IdVendor == id)
+                    {
+                        hasOffer = true;
+                        foreach (var purchase in listPurchase)
+                        {
+                            if (purchase.IdOffer == offer.IdOffer)
+                            { 
+                                hasPurchase= true;
+                                break;
+                            }
+                        }
+                    }
+                }
+
+                if (!hasPurchase)
+                {
+                    if (hasOffer)
+                    {
+                        foreach (var offer in listOffer)
+                        {
+                            if (offer.IdVendor == id)
+                            {
+                                _offerRepository.DeleteOffer(offer.IdOffer);
+                            }
+                        }
+                    }
+                    _vendorRepository.DeleteVendor(id);
+                    return RedirectToAction(nameof(Index));
+                }
+                else
+                {
+                    TempData["VendorErrorMessage"] = "This vendor is associated with an offer that has a purchase.Cannot delete!";
+                    return RedirectToAction("Delete",id);
+                }
             }
-            catch
+            catch (Exception ex)
             {
+                TempData["VendorErrorMessage"] =ex.Message;
                 return RedirectToAction("Delete",id);
             }
         }
